Ensure consistent indentation, using Xcode default soft tabs
There was a mix of hard and soft tab indenting in use; Xcode uses soft tabs of four spaces to indent code blocks, and so every tab in the Swift source is now replaced by four spaces.

This makes browsing the source code in other external editors or on GitHub significantly easier.